% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_interpolate_points.R
\name{aggregate_interpolate_points}
\alias{aggregate_interpolate_points}
\title{Interpolate and temporal aggregate DWD values}
\usage{
aggregate_interpolate_points(dataframe, coords, epsg, DateTime, infection,
  incubation, aim_variable, outputfile, trans_epsg = FALSE,
  co_variables = FALSE, procedure = c("ked", "ok", "idw"),
  progressbar = TRUE)
}
\arguments{
\item{dataframe}{dataframe containing the aim variable and if "ked" should be applied the covariables}

\item{coords}{vector containing names of the columns containing x and y coordinate values}

\item{epsg}{number of the EPSG code of the "coords" information}

\item{DateTime}{name of the DateTime column}

\item{infection}{duration of the assumed infection}

\item{incubation}{duration of the assumed incubation}

\item{aim_variable}{Character string with the name of the aim variable}

\item{outputfile}{SpatialPointsDataframe, SpatialGridDataFrame or raster which should be filled with predictions; requires covariables for "ked"}

\item{trans_epsg}{default = FALSE, number of the EPSG code the "coords" information should be transformed to}

\item{co_variables}{default = FALSE, vector of covariables if needed}

\item{procedure}{default = c("ked","ok","idw"); vector containing the interpolation technic to be used; the first method is used and if this does not work out, the second, and so on}

\item{progressbar}{default = TRUE; should a progressbar be generated?}
}
\value{
a dataframe or a SpatialPointsDataFrame containing information about DWD locations in Germany
}
\description{
This functions interpolates points based on a dataframe containing coordinates, an aim variable, an aim dataset and a parameter such as a date, by which the interpolations are divided
}
\examples{

# Download example data
shdat <- download_statewide_hourly_station_data(state = "Schleswig-Holstein", coord = TRUE)

shdat2 <- shdat \%>\%
    filter(DateTime < sort(unique(shdat$DateTime))[50])

example <- aggregate_interpolate_points(dataframe = shdat2,
                                       coords = c("lon","lat"),
                                       DateTime = "DateTime",
                                       infection=2,
                                       incubation=8,
                                       aim_variable ="Temperature",
                                       outputfile=c(1000,1000),
                                       co_variables = FALSE,
                                       procedure = c("ked","ok","idw"),
                                       epsg = 4326,
                                       trans_epsg = 25832)
plot(example[[1]])
}
\author{
Wolfgang Hamer
}
