% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_statistical_measures.R
\name{apply_statistical_measures}
\alias{apply_statistical_measures}
\title{Calculate statistical measures}
\usage{
apply_statistical_measures(dataframe, ObservedName = "Observed",
  PredictedName = "Predicted")
}
\arguments{
\item{dataframe}{dataframe containing Observed and Predicted variables as created by `loocv_machine_learner`}

\item{ObservedName}{Character string of the Observed column; default = "Observed"}
}
\value{
dataframe containing Accuracy, Sensitivity, Specificity, Precision and ROC AUC
}
\description{
This functions calculates statistical measures for dataframes containing Observed and Predicted factors
}
\author{
Wolfgang Hamer
}
