% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learner.R
\name{auto_machine_learner}
\alias{auto_machine_learner}
\title{Apply machine learning functions}
\usage{
auto_machine_learner(
  dataframe,
  aim_variable,
  co_variables,
  method = c("DT", "BDT", "RF")
)
}
\arguments{
\item{dataframe}{dataframe containing variables of interest}

\item{aim_variable}{Character string with the name of the aim variable}

\item{co_variables}{Character string with the name of the co-variables}

\item{method}{default = c("DT","BDT","RF"); which method should be used: DecisionTree, BoostedDecisionTree and/or RandomForest?}

\item{additionalparameters}{default = FALSE; list containing additional parameters for the ML procedures / e.g.: list(RF=list(ntree=10000))}
}
\value{
list containing the models
}
\description{
This functions applies one or several machine learning methods on a given dataset
}
\author{
Wolfgang Hamer
}
