% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_station_data.R
\name{download_alltime_hourly_station_data}
\alias{download_alltime_hourly_station_data}
\title{Download alltime hourly weather data of the DWD stations}
\usage{
download_alltime_hourly_station_data(station,
  parameter = c("temperature", "precipitation", "windspeed"),
  astbl = FALSE)
}
\arguments{
\item{station}{the station ID}

\item{parameter}{one ore multiple paramters of c("temperature", "humidity", "precipitation", "windspeed", "winddirection")}

\item{astbl}{default = FALSE; should the explort be a dataframe or a tibble}
}
\value{
a p value of the comparison between the selected and random points
}
\description{
Downloads historical and recent hourly data of the DWD stations in Germany
}
\examples{

# Select Location
mapview::mapview(get_all_dwd_locations(TRUE))

Fehmarn <- download_alltime_hourly_station_data(5516)
head(Fehmarn)

LeuchtturmKiel <- download_alltime_hourly_station_data(02961, parameter = "windspeed")
head(LeuchtturmKiel)
}
\author{
Wolfgang Hamer
}
