% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_station_data.R
\name{download_hourly_station_data}
\alias{download_hourly_station_data}
\title{Download hourly weather data of the DWD stations}
\usage{
download_hourly_station_data(station, parameter = c("temperature",
  "precipitation", "windspeed"), time = "recent", astbl = FALSE)
}
\arguments{
\item{station}{the station ID}

\item{parameter}{one ore multiple paramters of c("temperature", "humidity", "precipitation", "windspeed", "winddirection")}

\item{time}{either "recent" often updated data or "historical" data which go longer in the past}

\item{astbl}{default = FALSE; should the explort be a dataframe or a tibble}
}
\value{
a p value of the comparison between the selected and random points
}
\description{
Downloads hourly data of the DWD stations in Germany
}
\examples{

# Select Location
mapview::mapview(get_all_dwd_locations(TRUE))

Fehmarn <- download_hourly_station_data(5516)
head(Fehmarn)

LeuchtturmKiel <- download_hourly_station_data(02961, parameter = "windspeed")
head(LeuchtturmKiel)
}
\author{
Wolfgang Hamer
}
