% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_station_data.R
\name{download_statewide_hourly_station_data}
\alias{download_statewide_hourly_station_data}
\title{Download hourly weather data of the DWD stations of federal states}
\usage{
download_statewide_hourly_station_data(
  state,
  parameter = c("temperature", "precipitation", "windspeed"),
  time = "recent",
  coord = FALSE,
  savefile = FALSE
)
}
\arguments{
\item{state}{the Federal State (e.g. "Schleswig-Holstein")}

\item{parameter}{one ore multiple paramters of c("temperature","humidity","precipitation","windspeed","winddirection")}

\item{time}{either "recent" often updated data or "historical" data which go longer in the past}

\item{coord}{default = FALSE; should the explort contain coordinates or not}

\item{savefile}{default = FALSE; where should the file be saved as .csv file?}
}
\value{
a p value of the comparison between the selected and random points
}
\description{
Downloads hourly data of the DWD stations in one federal state in Germany
}
\examples{

shdat <- download_statewide_hourly_station_data(state = "Schleswig-Holstein", coord = TRUE)
head(shdat)
}
\author{
Wolfgang Hamer
}
