% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dwd_raster.R
\name{download_windspeed_dwd_raster}
\alias{download_windspeed_dwd_raster}
\title{Download multi-annual windspeed DWD rasters}
\usage{
download_windspeed_dwd_raster(parameter = "windspeed", period = "",
  heigth = "", crop = FALSE)
}
\arguments{
\item{parameter}{a character string defining the parameter to be downloaded ("e.g.: "windspeed")}

\item{period}{years which are combined in the mult annual datasets (e.g.: "1961-1990", "1981-2010","1991-2010", "1992-2015")}

\item{heigth}{the heigth of the observation (e.g. "10m")}

\item{crop}{Spatial Dataset of which an extent can be created which is used to crop the germany wide DWD dataset}
}
\value{
a raster dataset
}
\description{
This functions download multi-annual windspeed DWD rasters and crops them if desired
}
\author{
Wolfgang Hamer
}
