% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_add_date_time.R
\name{dwd_add_date_time}
\alias{dwd_add_date_time}
\title{Replaces DateTime column by Date (date) and Time (numeric) columns}
\usage{
dwd_add_date_time(dataframe, columnname = "DateTime")
}
\arguments{
\item{dataframe}{a dataframe}

\item{columnname}{default = "DateTime"; should contain values in the format "2017072602" for 2 oclock at the 26 th of Julya in 2017}
}
\value{
a dataframe like dataframe with two new columns
}
\description{
This functions replaces DateTime column by Date (date) and Time (numeric) columns
}
\examples{

locs <- get_dwd_locations(sp = TRUE)
mapview::mapview(locs)
}
\author{
Wolfgang Hamer
}
