% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_locations.R
\name{get_all_dwd_locations}
\alias{get_all_dwd_locations}
\title{Download all available DWD location data from the CDC Server}
\usage{
get_all_dwd_locations(sp = FALSE)
}
\arguments{
\item{sp}{default = FALSE; if TRUE returns not the plain data frame but a spatialised version}
}
\value{
a dataframe or a SpatialPointsDataFrame containing information about DWD locations in Germany
}
\description{
This functions downloads DWD location data from the CDC Server
}
\examples{

mapview::mapview(get_all_dwd_locations(TRUE))
}
\author{
Wolfgang Hamer
}
