% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dwd_locations.R
\name{get_dwd_locations}
\alias{get_dwd_locations}
\title{Download DWD location data from the CDC Server}
\usage{
get_dwd_locations(sp = FALSE, parameter = "temperature")
}
\arguments{
\item{sp}{default = FALSE; if TRUE returns not the plain data frame but a spatialised version}

\item{parameter}{default = "temperature"; should the "temperature" (and humidity), "precipitation" or "wind" station network be downloaded}
}
\value{
a dataframe or a SpatialPointsDataFrame containing information about DWD locations in Germany
}
\description{
This functions downloads DWD location data from the CDC Server
}
\examples{

locs <- get_dwd_locations(sp = TRUE)
mapview::mapview(locs)
}
\author{
Wolfgang Hamer
}
