% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfun.R
\name{large_ctu}
\alias{large_ctu}
\title{Apply ctu on large dataset}
\usage{
large_ctu(
  dataset,
  temp_column,
  date_column,
  start_date = "10-01",
  location_column = FALSE,
  vector = TRUE,
  TBD = 0,
  TP1D = 25,
  TP2D = 28,
  TCD = 40
)
}
\arguments{
\item{dataset}{a dataset}

\item{temp_column}{name of the  temperature column}

\item{date_column}{name of the  date column}

\item{start_date}{start date of the growing plant; defalut = "10-01" for October the 10th}

\item{location_column}{name of the  location column; defalut = FALSE}

\item{vector}{default = TRUE; boolean operator defining if a dataset with additional column or only the new column should be given out}

\item{TBD}{thermal base temperature; default = 0 for wheat}

\item{TP1D}{lower optimum temperature; default = 25 for wheat}

\item{TP2D}{upper optimum temperature; default = 28 for wheat}

\item{TCD}{thermal ceiling temperature; default = 40 for wheat}
}
\value{
a vector or raster file (depending on input) with the relative development rate based on temperature
}
\description{
This functions calculates Cumulative Thermal Units for large datasets using the daily thermal unit (dtu)
(Based on \url{https://www.researchgate.net/publication/281674392_Modeling_physiology_of_crop_development_growth_and_yield})
}
\examples{
# Download example data
shdat <- download_statewide_hourly_station_data(state = "Schleswig-Holstein", coord = TRUE)
shdat2 <- shdat \%>\% filter(DateTime > 1995093023)

shdat2 \%<>\% dplyr::mutate(Date = as.Date(substr(DateTime,1,8),"\%Y\%m\%d"))

shdat2 \%<>\% dplyr::mutate( CTU = large_ctu(dataset = shdat2,
                                           temp_column = "Temperature",
                                           date_column = "Date",
                                           start_date = "10-01",
                                           location_column = "ID"))
}
\author{
Wolfgang Hamer
}
