% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_in_CDC_folder.R
\name{list_files_in_CDC_folder}
\alias{list_files_in_CDC_folder}
\title{List files in CDC FTP folder}
\usage{
list_files_in_CDC_folder(path)
}
\arguments{
\item{path}{the path to be explored}
}
\value{
a vector with files stored in specific path
}
\description{
List files in CDC FTP folder
}
\examples{

list_files_in_CDC_folder("ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/hourly/air_temperature/historical/")
}
\author{
Wolfgang Hamer
}
