% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learner.R
\name{loocv_machine_learner}
\alias{loocv_machine_learner}
\title{Apply loocv machine learning functions}
\usage{
loocv_machine_learner(dataframe, aim_variable, co_variables,
  location = FALSE, method = c("DT", "BDT", "RF"),
  additionalparameters = FALSE, auto = FALSE)
}
\arguments{
\item{dataframe}{dataframe containing variables of interest}

\item{aim_variable}{Character string with the name of the aim variable}

\item{co_variables}{Character string with the name of the co-variables}

\item{location}{defalut = FALSE; Character string with the name of the location. If FALSE each observation is treated as unique location}

\item{method}{default = c("DT","BDT","RF"); which method should be used: DecisionTree, BoostedDecisionTree and/or RandomForest?}

\item{additionalparameters}{default = FALSE; list containing additional parameters for the ML procedures / e.g.: list(RF=list(ntree=10000))}

\item{auto}{default = FALSE; boolean operator defining if "machine_learner" or "auto_machine_learner" should be used}
}
\value{
list containing the models
}
\description{
This functions applies one or several machine learning methods on a given dataset and checks for the validity of the prediction
}
\author{
Wolfgang Hamer
}
