% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_fitter.R
\name{machine_fitter}
\alias{machine_fitter}
\title{Apply fitting of machine learning functions}
\usage{
machine_fitter(
  dataframe,
  aim_variable,
  co_variables,
  method = c("BDT", "RF"),
  splitper = 70
)
}
\arguments{
\item{dataframe}{dataframe containing variables of interest}

\item{aim_variable}{Character string with the name of the aim variable}

\item{co_variables}{Character string with the name of the co-variables}

\item{method}{default = c("DT","BDT","RF"); which method should be used: DecisionTree, BoostedDecisionTree and/or RandomForest?}

\item{splitper}{default = 70; percentage of dataset used for fitting the model}
}
\value{
list containing the models
}
\description{
This functions fits one or several machine learning methods on a given dataset
}
\author{
Wolfgang Hamer
}
