% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_predictor.R
\name{machine_predictor}
\alias{machine_predictor}
\title{Predict by raster stack and machine learning model}
\usage{
machine_predictor(
  rstack,
  mmodel,
  additionalRaster = FALSE,
  type = FALSE,
  index = FALSE
)
}
\arguments{
\item{rstack}{list containing an raster stack with covariables for prediction based on the mmodel}

\item{mmodel}{machine learning model}

\item{additionalRaster}{rasters that are identical in each time step and should be added to each rasterstack}

\item{type}{character string containing type of prediction (e.g. "prob" for probability); default to FALSE}

\item{index}{in case of type = "prob" the index of the parameter of which the probability shold be returned}
}
\value{
stack with one prediction for each element of the input list
}
\description{
This functions applies one machine learning model on a given rasterstack
}
\author{
Wolfgang Hamer
}
