% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfun.R
\name{one_year}
\alias{one_year}
\title{Give out one year}
\usage{
one_year(fromdate)
}
\arguments{
\item{fromdate}{name of the  temperature column}
}
\value{
a vector of one year following the date given
}
\description{
This function gives the sequence of one year following the date given
}
\examples{
one_year(as.Date("1996-10-01","\%Y-\%m-\%d"))
}
\author{
Wolfgang Hamer
}
