% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_input.R
\name{reduce_input}
\alias{reduce_input}
\title{Reduce the size of an given dataset}
\usage{
reduce_input(dataframe, DateTime, infection, incubation, event_dates)
}
\arguments{
\item{dataframe}{dataframe containing variables of interest which should be reduced}

\item{DateTime}{name of the DateTime column}

\item{infection}{duration of the assumed infection}

\item{incubation}{duration of the assumed incubation}

\item{event_dates}{Character string with the name of the aim variable}
}
\value{
reduced dataframe input
}
\description{
This functions reduces the size of an given dataset in respect to the weather data of "infection" days relevant for an infestation "incubation" days later
}
\author{
Wolfgang Hamer
}
