% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_rasterstacklist.R
\name{store_rasterstacklist}
\alias{store_rasterstacklist}
\title{Stores a list of raster stacks}
\usage{
store_rasterstacklist(rstacklist, pathfolder)
}
\arguments{
\item{rstacklist}{a list of raster stacks}

\item{pathfolder}{folder to which the lists rasterstacks should be exported}
}
\description{
This function stores a list of raster stacks
}
\author{
Wolfgang Hamer
}
