% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempfun.R
\name{tempfun}
\alias{tempfun}
\title{Calculate response of relative development rate to temperature}
\usage{
tempfun(TMP, TBD = 0, TP1D = 25, TP2D = 28, TCD = 40)
}
\arguments{
\item{TMP}{temperature in °C (vector or raster file)}

\item{TBD}{thermal base temperature; default = 0 for wheat}

\item{TP1D}{lower optimum temperature; default = 25 for wheat}

\item{TP2D}{upper optimum temperature; default = 28 for wheat}

\item{TCD}{thermal ceiling temperature; default = 40 for wheat}
}
\value{
a vector or raster file (depending on input) with the relative development rate based on temperature
}
\description{
This functions calculates response of relative development rate to temperature depending on crop parameters
(Based on \url{https://www.researchgate.net/publication/281674392_Modeling_physiology_of_crop_development_growth_and_yield})
}
\examples{
tempfun(10)
tempfun(c(23,12,23))
}
\author{
Wolfgang Hamer
}
