% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_rasterstacklist.R
\name{read_rasterstacklist}
\alias{read_rasterstacklist}
\title{Reads a list of raster stacks}
\usage{
read_rasterstacklist(pathfolder)
}
\arguments{
\item{pathfolder}{folder to which the lists rasterstacks should be exported}
}
\description{
This function reads a list of raster stacks as stored by store_rasterstacklist
}
\author{
Wolfgang Hamer
}
